requirements.txt
kagglehub
pandas
langchain-community
langchain-text-splitters
langchain-google-genai
python-dotenv
seaborn
matplotlib
numpy
transformers
gradio
tqdm
langchain
chromadb